pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
torch = "1.13.0"
transformers = "4.29.2"
accelerate = "0.19.0"
clip = "0.2.0"
torchvision = "0.14.0"
python-dotenv = "^1.0.1"
diffusers = "0.16.1"
openai = "^1.45.0"
opencv-python = "^4.10.0.84"
facenet-pytorch = "2.5.3"
numpy = "1.26.4"
matplotlib = "^3.9.2"
wandb = "^0.18.1"
mediapipe = "^0.10.15"
pandas = "^2.2.3"
google-generativeai = "^0.8.2"
huggingface-hub = "0.16.4"
pillow = "^11.0.0"
icecream = "^2.1.3"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pyproject-flake8 = "^7.0.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
